+++ local/backend.go
@@ -21,12 +21,12 @@ import (
 
 	"github.com/docker/docker/client"
 
+	"github.com/docker/compose-cli/api/backend"
 	"github.com/docker/compose-cli/api/compose"
 	"github.com/docker/compose-cli/api/containers"
 	"github.com/docker/compose-cli/api/resources"
 	"github.com/docker/compose-cli/api/secrets"
 	"github.com/docker/compose-cli/api/volumes"
-	"github.com/docker/compose-cli/backend"
 	"github.com/docker/compose-cli/context/cloud"
 	local_compose "github.com/docker/compose-cli/local/compose"
 )
